Make thumbnail updates when scrolling go from first to last, not reverse.
What we want to achieve here is:
1. Once user scrolls, we want to start showing thumbnails for items
visible after scroll ASAP.
2. We want to show them from beginning to end, not the other way around
as was in original ComicsReader.
3. But first line may be almost completely obscured, so no actual thumbs
is visible on it (only text titles), so we want to skip them.
4. Or first line may be only partially obscure, and small part of thumb
may be visible. In this case, we want to start with fully visible lines,
but get back to first partial line in the end.